requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
idna==2.10
joblib==0.17.0
murmurhash==1.0.2
nltk==3.5
numpy==1.19.2
plac==1.1.3
preshed==3.0.2
python-Levenshtein==0.12.0
regex==2020.10.15
requests==2.24.0
six==1.15.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.50.2
treetaggerwrapper==2.3
urllib3==1.25.10
wasabi==0.8.0
